Split stuff into lib and main
