+++ .gitignore
@@ -1,2 +1,3 @@
 /target
 .cargo/config.toml
+/wasm
